# openapi/openapi.yml
openapi: 3.0.0
info:
  title: openapi.yml
  version: '1.0'
paths:
  /items:
    get:
      summary: Your GET endpoint
      tags:
        - items
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    description: アイテム一覧
                    items:
                      $ref: '#/components/schemas/item'
                required:
                  - items
      operationId: get-items
      description: アイテム一覧を取得します
      parameters:
        - schema:
            type: integer
            minimum: 0
          in: query
          name: $top
          description: 取得数
    post:
      summary: ''
      operationId: post-items
      tags:
        - items
      description: アイテムを一件登録します
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/item'
      responses:
        '201':
          description: Created
      security:
        - Authorization: []
    parameters: []
  '/items/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Your GET endpoint
      tags:
        - items
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/item'
                required:
                  - item
        '404':
          description: Not Found
      operationId: get-item
    put:
      summary: ''
      operationId: put-item
      responses:
        '204':
          description: No Content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/item'
      security:
        - Authorization-Admin: []
      tags:
        - items
    delete:
      summary: ''
      operationId: delete-item
      responses:
        '204':
          description: No Content
      security:
        - Authorization-Admin: []
components:
  schemas:
    item:
      title: item
      x-stoplight:
        id: ly7bu2x9amqta
      type: object
      description: アイテム
      properties:
        id:
          type: string
        name:
          type: string
          description: アイテム名
        description:
          type: string
          description: アイテム説明
        price:
          type: integer
          minimum: 0
          description: 価格
      required:
        - id
        - name
        - price
  securitySchemes:
    Authorization:
      name: Authorization
      type: apiKey
      in: header
    Authorization-Admin:
      name: Authorization-Admin
      type: apiKey
      in: header